# docs/swagger.yaml
definitions:
  api.CheckStatusReq:
    properties:
      name:
        type: string
      version:
        type: string
    type: object
  api.CheckStatusResp:
    properties:
      status:
        type: boolean
    type: object
  api.Resp:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  api.SetStatusReq:
    properties:
      password:
        type: string
      programs_name:
        type: string
      status:
        type: boolean
      username:
        type: string
      version:
        type: string
    type: object
info:
  contact: {}
paths:
  /change/checkStatus:
    post:
      consumes:
      - application/json
      description: 从 change_miniprograms 表中查询状态
      parameters:
      - description: 查询参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.CheckStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Resp'
            - properties:
                data:
                  $ref: '#/definitions/api.CheckStatusResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Resp'
      summary: 查询变更项目状态
      tags:
      - change
  /change/setStatus:
    put:
      consumes:
      - application/json
      description: 更新 change_miniprograms 表中的状态
      parameters:
      - description: 设置参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.SetStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Resp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Resp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Resp'
      summary: 设置变更项目状态
      tags:
      - change
  /checkStatus:
    post:
      consumes:
      - application/json
      description: 根据项目名称查询当前启用状态
      parameters:
      - description: 查询参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.CheckStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Resp'
            - properties:
                data:
                  $ref: '#/definitions/api.CheckStatusResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Resp'
      summary: 查询项目状态
      tags:
      - base
  /setStatus:
    put:
      consumes:
      - application/json
      description: 创建或更新项目状态（需要用户名密码）
      parameters:
      - description: 设置参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.SetStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Resp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Resp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Resp'
      summary: 设置项目状态
      tags:
      - base
swagger: "2.0"